+++ lib/constants/strings.dart
@@ -14,7 +14,7 @@ final String kDeviceIdKey = 'kDeviceIdKey';
 const String kDifficultyKey = 'kDifficultyKey';
 const String kGameState = 'kGameState';
 
-const String shareIncomplete = 'You can\'t share a furdle that hasn\'t been solved yet!';
+
 const String copiedToClipBoard = 'Score copied to clipboard';
 const String gameAlreadyPlayed = 'You have already played the puzzle';
 const String furdleCracked = 'You cracked the Furdle of the Day!\n\nNext Furdle in';