+++ lib/provider/hive_storage_provider.dart
@@ -1,5 +1,4 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:furdle/service/storage_service.dart';
 import 'package:furdle/service/hive_storage_service.dart';
 
 final hiveStorageServiceProvider = Provider<HiveStorageService>((ref) {