+++ lib/utils/utility.dart
@@ -1,8 +1,6 @@
-import 'package:flutter/material.dart';
 import 'package:furdle/constants/const.dart';
 import 'package:furdle/provider/game_state_notifier.dart';
 import 'package:furdle/state/game_state.dart';
-import 'package:furdle/utils/widgets.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 /// State Color for either furdle or Keyboard
@@ -16,17 +14,8 @@ class Utility {
     );
   }
 
-  static void showMessage(context, message,
-      {Duration? duration = const Duration(milliseconds: 1500), EdgeInsetsGeometry? margin}) {
-    ScaffoldMessenger.of(context)
-        .showSnackBar(snackBar(message: '$message', duration: duration!, margin: margin));
-  }
-
   static String generateFurdleGrid(GameState state) {
     final isPuzzleCracked = state.status == GameStatus.win;
-    if (!isPuzzleCracked) {
-      return 'You can\'t share a furdle that hasn\'t been solved yet!';
-    }
     final int attempts = isPuzzleCracked ? state.row + 1 : 0;
     String generatedFurdle = '#${state.id} $attempts/${state.size.height.toInt()}\n\n';
     for (int i = 0; i < state.size.height; i++) {