+++ changelog.md
@@ -1,4 +1,13 @@
-### v0.4.1+15 (Unpublished)
+### v0.4.2+16 (Unpublished)
+
+- Make Keyboard taps more accurate
+
+### v0.4.2+16 Sep 27, 2025
+
+- Improve the Keyboard Layout for better usability.
+- Update notification should launch playstore.
+- Update UI of title bar.
+- Show Share button only on completion of Daily challenge.
 
 ### v0.4.1+15 Sep 27, 2025